refactor: expose CSharpBaseExecutorProxy via _internal
